updated contract to included less input on intializing of clone contract
