--- .env.local
@@ -1,2 +1,0 @@
-ALCHEMY_DEV_KEY=https://eth-rinkeby.alchemyapi.io/v2/5RR-EGA3lFlRlT_TZ13qhGF7SzftrO5B
-ADDRESS_KEY=3ab0a26a615d18193eae0d2e394dc355f7b7a1188ca4a134dd1e4a02ace0b905
