+++ hardhat.config.js
@@ -1,5 +1,11 @@
 require("@nomiclabs/hardhat-waffle");
+require("@nomiclabs/hardhat-etherscan");
+require("hardhat-gas-reporter");
 require("dotenv").config();
+
+console.log(process.env.ADDRESS_KEY);
+
+// const ETHERSCAN_API = process.env.ETHERSCAN_API;
 // This is a sample Hardhat task. To learn how to create your own go to
 // https://hardhat.org/guides/create-task.html
 task("accounts", "Prints the list of accounts", async (taskArgs, hre) => {
@@ -18,12 +24,18 @@ task("accounts", "Prints the list of accounts", async (taskArgs, hre) => {
  */
 module.exports = {
   solidity: "0.8.4",
+  defaultNetwork: "rinkeby",
   networks: {
-    hardhat: {
-    },
-    // rinkeby: {
-    //   url: `${process.env.ALCHEMY_DEV_KEY}`,
-    //   accounts: [process.env.ADDRESS_KEY]
-    // }
+    hardhat: {},
+    rinkeby: {
+      url: `https://eth-rinkeby.alchemyapi.io/v2/${process.env.ALCHEMY_DEV_KEY}`,
+      accounts: [process.env.ADDRESS_KEY]
+    }
+  },
+  etherscan: {
+    apiKey: process.env.ETHERSCAN_API
+  },
+  gasReporter: {
+    enabled: true
   }
 };
