+++ package.json
@@ -2,14 +2,16 @@
   "name": "hardhat-project",
   "devDependencies": {
     "@nomiclabs/hardhat-ethers": "^2.0.5",
+    "@nomiclabs/hardhat-etherscan": "^3.0.3",
     "@nomiclabs/hardhat-waffle": "^2.0.3",
     "chai": "^4.3.6",
+    "dotenv": "^16.0.0",
     "ethereum-waffle": "^3.4.0",
     "ethers": "^5.6.1",
     "hardhat": "^2.9.1"
   },
   "dependencies": {
     "@openzeppelin/contracts": "^4.5.0",
-    "dotenv": "^16.0.0"
+    "hardhat-gas-reporter": "^1.0.8"
   }
 }
