+++ werkJson.json
@@ -1,0 +1,12 @@
+{
+    "name": "Nick Gonzalez",
+    "scope": "Make me a mf meme, please.",
+    "expertise": "Engineering, Memes",
+    "length (days)": "31",
+    "deadline": "4/01/22",
+    "start": "5/01/22",
+    "end": "5/31/22",
+    "method": "ETH",
+    "reward": "0.7",
+    "client": "Peng"
+}
